requirements.txt
pandas==1.3.4
google-ads==14.1.0
dateparser==1.1.0